feat(IntegrationTest): 쿠키 생성 책임 밖으로 빠지도록 수정
