+++ src/test/java/roomescape/model/IntegrationTest.java
@@ -6,11 +6,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.ResponseCookie;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
-import roomescape.auth.domain.Token;
-import roomescape.auth.provider.CookieProvider;
 import roomescape.auth.provider.model.TokenProvider;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -22,16 +19,11 @@ public abstract class IntegrationTest {
     @LocalServerPort
     private int port;
     @Autowired
-    private TokenProvider tokenProvider;
-
-    protected Token token;
-    protected ResponseCookie cookie;
+    protected TokenProvider tokenProvider;
 
     @BeforeEach
     void init() {
         RestAssured.port = port;
-        this.token = tokenProvider.getAccessToken(1);
-        this.cookie = CookieProvider.setCookieFrom(token);
     }
 
 }
