+++ src/test/java/roomescape/reservation/integration/ReservationIntegrationTest.java
@@ -6,6 +6,9 @@ import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.http.ResponseCookie;
+import roomescape.auth.domain.Token;
+import roomescape.auth.provider.CookieProvider;
 import roomescape.model.IntegrationTest;
 import roomescape.reservation.dto.ReservationRequest;
 
@@ -14,6 +17,9 @@ class ReservationIntegrationTest extends IntegrationTest {
     @Test
     @DisplayName("정상적인 요청에 대하여 예약을 정상적으로 등록, 조회, 삭제한다.")
     void adminReservationPageWork() {
+        Token token = tokenProvider.getAccessToken(1);
+        ResponseCookie cookie = CookieProvider.setCookieFrom(token);
+
         ReservationRequest reservationRequest = new ReservationRequest(TODAY.plusDays(1), 1L, 1L);
 
         RestAssured.given().log().all()
@@ -45,6 +51,9 @@ class ReservationIntegrationTest extends IntegrationTest {
     @Test
     @DisplayName("예약을 요청시 존재하지 않은 예약 시간의 id일 경우 예외가 발생한다.")
     void notExistTime() {
+        Token token = tokenProvider.getAccessToken(1);
+        ResponseCookie cookie = CookieProvider.setCookieFrom(token);
+
         ReservationRequest reservationRequest = new ReservationRequest(TODAY, 0L, 1L);
 
         RestAssured.given().log().all()
